[feat(1/3)] Support DeepSeek V3.2 Exp (weight/config/misc) (merge request !1420)
Squash merge branch 'dsv3.2-misc' into 'master'
DeepSeek V3.2 Exp模型的权重加载、配置解析、KV Cache管理

TAPD: --story=127796837

+++ src/ksana_llm/service/inference_engine.cpp
@@ -197,8 +197,8 @@ Status InferenceEngine::Initialize() {
   status = env->GetProfilerConfig(profiler_config);
   Singleton<Profiler>::GetInstance()->InitMetrics(profiler_config);
 
-  size_t max_batch_size = (size_t)runtime_config.max_batch_size;
-  size_t max_vocab_size = (size_t)model_config.vocab_size;
+  size_t max_batch_size = runtime_config.max_batch_size;
+  size_t max_vocab_size = model_config.vocab_size;
 
   status = LoadOperatorOptimization(model_config);
   if (!status.OK()) {
